pyproject.toml
[project]
name = "aygentx-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "fastapi>=0.135.3",
    "mcp>=1.27.0",
    "openai>=2.31.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
